+++ CHANGELOG.md
@@ -7,8 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- idk lol
-  - 
+- IntelliSense
+  - Tag scope detector
 
 ## [0.3]